Proyecto completo

+++ src/app/components/article/article.component.ts
@@ -1,5 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { Router, ActivatedRoute, Params } from '@angular/router';
+import swal from 'sweetalert';
 import { Global } from '../../services/global';
 import { ArticleService } from '../../services/article.service';
 import { Article } from '../../models/article';
@@ -46,15 +47,37 @@ export class ArticleComponent implements OnInit {
   } 
 
   delete(id){
-    this._articleService.delete(id).subscribe(
-      response => {
+
+    swal({
+      title: "¿Estas seguro?",
+      text: "Una vez borrado el articulo no podràs recuperarlo!",
+      icon: "warning",
+      buttons: [true, true],
+      dangerMode: true
+    })
+    .then((willDelete) => {
+      if (willDelete) {
+        this._articleService.delete(id).subscribe(
+          response => {
+        swal("El articulo ha sido borrado!", {
+          icon: "success",
+        });
         this._router.navigate(['/blog']);
       },
       error => {
-       console.log(error);
-        this._router.navigate(['/blog']);
+        console.log(error);
+         this._router.navigate(['/blog']);
+       }
+     );
+
+      } else {
+        swal("Tranquilo, nada se ha borrado");
       }
-    );
+    });
+  
+
+ 
+    
 
    
   }
